Qt: Don't select and scroll to packets twice
PacketList::goToPacket calls cf_goto_frame, which only succeeds
if packet_list_select_row_from_data succeeds. The latter
function already clears and selects the current row and scrolls
to it at the center. We don't need to find the row number a
second time, select and scroll to it again. Do select the hfid
if we're going to a different one (e.g., selected from the
expert info dialog - note that 0 is a valid hfid in 4.2 and
earlier, but not anymore.)

Also remove a comment that hasn't been true for years.