Qt: Limit on how often mouse moves update the RTP Player graph
The work done when updating the graph from moving the mouse
calls QCP plottableAt() multiple times. That calls pointDistance(),
which is particularly slow on graphs that have many lines (like
the audio plots), as it computes the distance between each line on
the graph to the point.

That's time-consuming because it doesn't rule out any line, because
line segments can pass close to the test point even if the two endpoints
of the line aren't particularly close.

While on my tests on Windows, I can't get a QMouseEvent to be issued
any more often than 10 ms, on Linux with mouse movement over the
graph I can get a QMouseEvent every 25 μs (at least, if the ponderous
calculations are disabled). As mouse tracking is on, this means that
the GUI becomes unresponsive whenever the mouse is moved over the graph,
constantly calculating distances between points on the graph where the
mouse was in the recent past and all the lines on the graph.

Set a QTimer and keep the actual calculations to some kind of sane
interval, just remembering the latest position if the timer is active.