epan: Remove subnum from frame_data
The RRC dissector is the only one which uses subnum, and fd->subnum
always contains exactly the same information as the cur_tb member
of the FP proto data (which is what the UMTS MAC and RLC dissectors
use instead.)

Switch the RRC dissector to use it as well, and remove subnum from
frame data. Move the new TCP analysis override member (which
probably can be a bitfield as well) near the bitfields in the
frame_data struct to replace it, to save space when allocating the
struct.