Qt: Scale IO Graph bar widths to the interval
This prevents bars from overlapping when intervals smaller than 1 sec
are used, and prevents huge gaps when larger intervals are used.

The default QCustomPlot width is 0.75 of the x-coordinate axis, which
creates a gap when the interval is 1 s. We could multiply by 0.75
or some other constant factor to have that effect.

Related to #17422, as it makes it easier to have the pen color for
the bars the same as the brush color.

Related to #13682, because bars are nigh useless on smaller
interals without scaling the width.