ICMP: fix minimum original datagram length
RFC 4884 section 3 requires the "original datagram" field be at minimum
128 bytes, with zero padding if the original datagram was shorter.  (And
this is in fact how routers on the internet behave in generating ICMP
messages with MPLS info.)

The ICMP dissector didn't implement this limit and ended up decoding the
zero padding as multi-part extension.  Fix by making 128 be the bottom
cap of padding to skip.