expert: register ei_unreassembled with proto_unreassembled.
We have two pseudo-protocols, not one, and proto_unreassembled is
intended for "Wireshark didn't reassemble this, so we ran past the end
of the first segment" type errors.  Have two separate ei_register_info[]
arrays, one for errors deemed "malformed" and one for errors deemed
"unreassembled", and register them with the appropriate protocols.