I/O Graphs: Squeeze some bytes out of io_graph_item_t
The min, max, and total members vary depending upon the hfinfo
associated with the graph (or aren't used at all), so make
them unions.

We don't need separate float and double members; floats can be
represented as doubles (and that's how we graph them anyway.)

For the integer types, use doubles for the totals (as we've
been doing; the integer total has been set but unused) because
of possible overflow from summing many 64 bit integer fields.
We only use the total when graphing (where it is converted to
a double anyway, so we don't care about any lost precision
for unsigned 64 bit integers larger than 2^53.) For the min and
max use 64 bit integers because of that possible precision loss
(so we always find the frame with the max and min value.)

Have separate members for the min and max frame in the interval,
so that we don't have to retap in order to find the max frame
if we previously were looking for the min, and vice versa. (In
practice, we were not retapping, so if we switched graph types
the frame numbers were wrong.)

This reduces sizeof(io_graph_item_t) from 152 bytes to 88 bytes
on a system with 64 bit time_t.

LOAD graphs still require retapping, since the value stored in
item->time_tot for a LOAD graph is not the same as that stored
in other graphs types. (Note however retapping doesn't occur
when changing graph types, see above, and also LOAD graphs are
broken since eb4e2cca69c5057995ff7.) We might want to add in
some bytes later so that the value for LOAD graphs can be stored
on the initial tap pass alongside the other values.