Qt: Treat IO Graphs with lines and data points as line graphs
Commit edf48633bbf5d8057e1929ef997d78eaad01351f added some
styles that have both lines and symbols for each data point
(a scatter style). Treat those as line graphs for the purposes
of displaying zero values instead of as scatter plots, so change
the test for if a graph is a scatter plot from "if the scatter
style is not none" to "if the line style is not none" to reflect that.

(Cf.
https://www.wireshark.org/docs/wsug_html_chunked/ChStatIOGraphs.html#ChStatIOGraphsMissingValues
"For scatter plots it is assumed that zero values indicate missing
data, and those values are omitted. Zero values are shown in line
graphs, and bar charts." )