IPPUSB: Avoid (mostly harmless) string overread
The result of tvb_get_ptr and the CHUNKED_END array are not
null terminated (but are guaranteed to be CHUNK_LENGTH_MIN),
so memcmp is more appropriate than strncmp, as the later will
call strlen, which goes outside the desired range. It's a read,
not a write, and it will be thrown away because it will always
compare larger than CHUNK_LENGTH_MIN, but there's no reason to
do it. It also, at least on some versions of gcc and some compiler
options, causes a stringop-overread warning that becomes an error.

For good practice, replace the tvb_get_ptr + memcmp call with
tvb_memeql.