Qt: Recalculate IO Graph when switching plottable
When switching the plottable from a QCPGraph to QCPBars or vice
versa, we have to recalculate the graph to insert the data.
If the plottable is the same but changes styles, we only need
a replot (which we do in the calling function.)

This was exposed by 81c7238776c3da437289c884c311373624a11b24,
which removed the automatic recalculation in nearly all
circumstances.