UAT: Handle TRUE and FALSE regardless of case
Handle "TRUE" or "FALSE" with any case for a UAT boolean.
Continue to write it as TRUE or FALSE in the files, for
backwards compatibility for versions before 4.4.
Note that the IO Graphs have their own version of this, from
when a UAT wasn't used (pre-2.6), though we can probably
stop writing "Enabled" and "Disabled" in that case (but
continue reading them for a while.)

Follow up to f3311c18403387bfcaed9c4d7798bdca663b67ed