Qt: Only retap when making an IO Graph visible if necessary
We don't retap when changing parameters of an IO Graph if it's
not visible. Remember whether or not we made a change that requires
retapping, and retap upon enabling a graph only if we need to.

As part of this, don't perform tapping calculations if graphs are
disabled, but mark the graph in need of a retap.

An alternative is to do the calculations anyway, and clear pending
retaps when the taps are reset. (For this dialog, the taps are
reset if something else retaps, or if the capture file is closed,
but in the latter case the tap listener is removed and it can never
be retapped again anyway.)

Fix #15822