Qt: Enable min/max button on GeometryStateDialogs by default
GeometryStateDialog is used when we want to remember user-set
geometry for a window. If we want the user to be able to control
the geometry, we probably want to allow minimization and maximization.

Pass Qt::Window as the default WindowFlag to GeometryStateDialog
(which is what WiresharkDialogs already do). This on most platforms
defaults to adding minimize and maximize buttons. (On some platforms,
like recent GNOME, we're at the mercy of various settings like

gsettings get org.gnome.desktop.wm.preferences button-layout

but the user can access minimize and maximize actions via right-clicking
the window title bar. Unfortunately it won't stop user confusion.)

Fix #18980. Part of #12566.