Qt: Fix IO Graph startOffset when packets aren't added
We don't always add certain zero points to the graph, even moreso
after hasItemToShow was introduced.

For currentOffset (which is used to determine start_time_ of the
entire IOGraphDialog, not just one IOGraph), we want the necessary
conversion between relative and absolute time so that absolute time
of day works for CSV output (#13717) and for packet selection with
Time of Day (#10922) when some items aren't added to the graph.