Qt: Work around Qt6 dataChanged pessimization
A change in Qt 6.0 makes QAbstractItemView::dataChanged, when called
with multiple indices, determine exactly how much of the viewport
rectangle is covered by the changed indices and only update that much
of the viewport instead of always updating the entire viewport. This
reduces CPU load when multiple indices, but not enough to span the
entire viewport, are updated at once.
( https://codereview.qt-project.org/c/qt/qtbase/+/285280
  https://bugreports.qt.io/browse/QTBUG-58580
)

Unfortunately, if the number of indices is very large (much larger
than the total number of indices that can possibly be displayed in
the viewport), computing the union of the intersecting rectangle takes
orders of magnitude longer than unconditionally updating the whole
viewport. ( https://bugreports.qt.io/browse/QTBUG-124173 )

When the packetListModel wants to indicate that all packets may have
changed, this makes dataChanged scale linearly with the total number of
rows/frames, whereas updating the viewport only scaled with the size of
the viewport but was unaffected by undisplayed packets. The calculation
is unnecessary because we know the entire viewport is affected. On a
sample modern workstation, measurements showed dataChanged taking 1 s
with 1.4 M packets and 9 s with 12 M packets. This signal can be issued
quite often, e.g. if IP address lookup (DNS or MaxMindIP) is enabled, it
can be issued once per second at the packet_data_timer expiration for
some period of time as idle dissection occurs when opening the file.

Instead of issuing dataChanged, we can issue layoutAboutToBeChanged()
(in practice a no-op) and layoutChanged(), which causes the QTreeView
to clear all information about its view items (d->viewItems.clear() in
QTreeView::doItemsLayout), but without clearing the current and selected
row (unlike [begin|end]ResetModel.) This takes the same 5-8 ms in the
same test procedure as above in both the 1.4 M and 12 M packet tests.

Improves #19486 considerably on Qt 6.