Clear queued rescans when opening a file
redissection_queued shouldn't be set when opening a file, but
make sure it's cleared. Otherwise things can enter a state where
the filter can't be changed or a redissection performed. cf_close
already checks the read lock and aborts if it's set, but it doesn't
hurt to clear it for a newly opened file.

When doing a cf_read, if there's a queued redissection at the
start of cf_read (which also shouldn't happen for similar reasons),
clear it for the same reason as clearing it at the start of
rescan_packets. The dissection in cf_read will satisfy any
pending request (and use the new filter.)

There's currently an issue with requesting a rescan (filtering,
redissection) when in the middle of retapping or other functions
that hold the read lock.