Qt: Use nstime_t in the rest of IO Graph
There's a question of how much precision to display in the
tooltip and to print in a CSV when using Time of Day / absolute
times. Since we don't align the bucket starts to the absolute times
(i.e., start 1 s buckets on integer seconds) but keep the buckets
as they are on the relative time, it should be related to the
precision of the absolute time offset. (The precision of the
absolute time for each frame does not have to be identical, at
least for pcapng, though.) For now we have some false precision
similar to other nstime timestamps (cf. 15579)

Related to #13682