ieee80211: Mark TPC Report link margin as reserved in certain frames
In accordance with IEEE Std 802.11-2012 and later, mark the TPC
Report link margin field as reserved in Beacon and Probe Response
frames.

Fix a few spellings. Add dBm units as appropriate to a few of
the tagged fields (most the corresponding fixed fields for the same
elements already have them.)

Fix #19751