pcapng: fix writing hash options
Since compute_block_option_size() and our option writing routines
already add needed padding, we don't need to compute the size
with padding in pcapng_compute_packet_hash_option_size(). That allows
us to use it both when computing the option size for the total
block size, and in the writing routine, for consistency.

Fix some errors where the type octet was not being included in
the bytes written for both one type of packet verdict, and for
packet hash. Also an error where the packet verdict option was
written with the queue option type.

Fix a possible buffer overrun for the fixed length hash types
when the hash length is wrong. We have a few different options
about how to handle hashes with bogus length - omit, write what
we have, truncate (if we have extra bytes) or err. Add some
comments regarding the options; for now, choose reporting an
error.

Fix #19766