Windows: Add include depended upon packets checkbox
Non-Windows already has this checkbox.

The vast majority of the time users do want dependent packets
so that reassembly, etc. works. Occasionally they don't (e.g.,
perhaps to exclude retransmitted segments or duplicates), but
the primary use of this checkbox is helping users understand why
extra packets are included so they don't think it's in error.

We might want to get rid of the special Windows dialogs. They
use a deprecated API, but the newer API is more difficult to customize.
The Qt common dialogs are better than the GTK+ file dialogs at the
time we started to use the Windows dialogs.

Follow up to dcc9cbffefe0598eadec11cbc1869943a26219c5
Fix #19772