http: Fix buffer overflow, use after free in HTTP Range
Don't blindly add to the start of a string without checking
if the proper prefix is there. Use strchr instead of strtok,
so as not to modify the string, which removes a strdup call.
Bounds check.

Use ws_strtou64, which has the advantage, unlike strtoul, of
treating a initial hyphen-minus as failure instead of a unary-minus
and applying the standard unsigned integer wraparound rules. We
want to handle suffix-ranges correctly.

Expand the comments about the edge cases that don't work.

If top_of_list is NULL after deleting requests, we definitely still
need to assign conv_data->req_list to it (i.e., NULL). NULL is an
empty GSList* - not assigning it to NULL means having it point to
freed memory and leads to use-after-free.

There are still some memory leaks to fix.

Fix #19779